refactor bindings into own file
